fab/outora-library/pyproject.toml
[project]
name = "outora-library"
version = "0.1.0"
description = "3D Outora Library asset pipeline and generator"
requires-python = ">=3.10"
readme = "README.md"
dependencies = [
    "Pillow",
]

[project.optional-dependencies]
dev = [
    "fake-bpy-module-latest",
    "ruff",
    "pytest",
    "black"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/outora_library"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []
